validate arguments

+++ fml/lib/FML/Command/Admin/rmdomain.pm
@@ -4,7 +4,7 @@
 #   All rights reserved. This program is free software; you can
 #   redistribute it and/or modify it under the same terms as Perl itself.
 #
-# $FML: rmdomain.pm,v 1.2 2003/11/16 11:53:18 fukachan Exp $
+# $FML: rmdomain.pm,v 1.3 2003/11/16 12:23:47 fukachan Exp $
 #
 
 package FML::Command::Admin::rmdomain;
@@ -58,9 +58,16 @@ sub process
     my $canon_argv = $command_args->{ canon_argv };
     my $domain     = $canon_argv->{ ml_name };
 
-    use FML::HomeDir::Prefix;
-    my $ml_home_prefix = new FML::HomeDir::Prefix $curproc;
-    $ml_home_prefix->delete($domain);
+    if ($domain) {
+	use FML::HomeDir::Prefix;
+	my $ml_home_prefix = new FML::HomeDir::Prefix $curproc;
+	$ml_home_prefix->delete($domain);
+    }
+    else {
+	my $error = 'no argument';
+	$curproc->logerror($error);
+	croak($error);
+    }
 }
